refactor: move dashboards into smaller components
